AFS_Summarizing_Assistant_Scrap_FuncApp/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai==1.3.5
requests==2.28.2